requirements.txt
asgiref==3.3.4
Django==3.2
gunicorn==20.1.0
mysql==0.0.2
mysqlclient==2.0.3
numpy==1.20.2
Pillow==8.2.0
pytesseract==0.3.7
pytz==2021.1
sqlparse==0.4.1
tesserocr==2.5.1
